# docs/swagger.yaml
basePath: /api/v1
definitions:
  addBuyMethod.EstelamOut:
    properties:
      _from:
        type: string
      _id:
        type: string
      _key:
        type: string
      _rev:
        type: string
      _to:
        type: string
      codeForSupplier:
        type: string
      createdAt:
        type: integer
      oneMonthPrice:
        type: boolean
      price:
        type: boolean
      show:
        type: boolean
      threeMonthPrice:
        type: boolean
      twoMonthPrice:
        type: boolean
      variant:
        type: string
    type: object
...
  addBuyMethod.PriceOut:
    properties:
      _from:
        type: string
      _id:
        type: string
      _key:
        type: string
      _rev:
        type: string
      _to:
        type: string
      codeForSupplier:
        type: string
      createdAt:
        type: integer
      oneMonthPrice:
        type: integer
      price:
        type: integer
      priceRepetition:
        type: integer
      show:
        type: boolean
      threeMonthPrice:
        type: integer
      totalNumber:
        type: integer
      totalNumberInCart:
        type: integer
      twoMonthPrice:
        type: integer
      variant:
        type: string
    type: object
...
  cart.CartOut:
    properties:
      _id:
        type: string
      _key:
        type: string
      _rev:
        type: string
      PricingType:
        type: string
      checkCommissionPercent:
        type: number
      commissionPercent:
        type: number
      createdAt:
        type: integer
      number:
        type: integer
      priceId:
        type: string
      pricePerNumber:
        type: integer
      productId:
        type: string
      productImageUrl:
        type: string
      productTitle:
        type: string
      supplierKey:
        type: string
      userKey:
        type: string
    type: object
...
  paymentAndWallet.PaymentByImage:
    properties:
      imageUrl:
        type: string
      includeTransportation:
        type: boolean
      orderKey:
        type: string
      overwritePaymentKey:
        type: boolean
      type:
        type: string
    type: object
  paymentAndWallet.addToWallet:
    properties:
      amount:
        type: integer
    type: object
  paymentAndWallet.checkByImage:
    properties:
      amount:
        type: integer
      checkNumber:
        type: string
      imageUrl:
        type: string
      orderKey:
        type: string
      status:
        type: string
      type:
        type: string
    type: object
  paymentAndWallet.filter:
    properties:
      checkNumber:
        type: string
      orderKey:
        type: string
      payerKey:
        type: string
      shaparakRefId:
        type: string
      status:
        type: string
      txType:
        type: string
    type: object
  paymentAndWallet.getSupplierConfirmationResponse:
    properties:
      cart:
        $ref: '#/definitions/cart.CartOut'
      info:
        $ref: '#/definitions/paymentAndWallet.supplierInfoForConfirmationOut'
    type: object
  paymentAndWallet.paymentOut:
    properties:
      _id:
        type: string
      _key:
        type: string
      _rev:
        type: string
      ShaparakRefId:
        type: string
      amount:
        type: integer
      cardHolder:
        type: string
      checkNumber:
        type: string
      createdAt:
        type: integer
      imageUrl:
        type: string
      includeTransportation:
        type: boolean
      orderKey:
        type: string
      payerKey:
        type: string
      status:
        type: string
      transId:
        type: string
      txType:
        type: string
    type: object
  paymentAndWallet.rejectionPoolItemOut:
    properties:
      _id:
        type: string
      _key:
        type: string
      _rev:
        type: string
      createdAt:
        type: integer
      number:
        type: integer
      paymentKey:
        type: string
      price:
        type: integer
      productId:
        type: string
      productImageUrl:
        type: string
      productTitle:
        type: string
      rejectBy:
        type: string
      sendInfoKey:
        type: string
      status:
        type: string
      txType:
        type: string
      userKey:
        type: string
    type: object
  paymentAndWallet.supplierInfoForConfirmationOut:
    properties:
      _id:
        type: string
      _key:
        type: string
      _rev:
        type: string
      cartKey:
        type: string
      orderKey:
        type: string
      supplierKey:
        type: string
    type: object
  paymentAndWallet.supplierPageResponse:
    properties:
      history:
        items:
          $ref: '#/definitions/paymentAndWallet.supplierWalletOut'
        type: array
      totalIn:
        type: integer
      totalOut:
        type: integer
      walletAmount:
        type: integer
    type: object
  paymentAndWallet.supplierWalletOut:
    properties:
      _id:
        type: string
      _key:
        type: string
      _rev:
        type: string
      amount:
        type: integer
      createdAt:
        type: integer
      income:
        type: boolean
      paymentKey:
        type: string
      supplierKey:
        type: string
      txStatus:
        type: string
      txType:
        type: string
    type: object
...
  products.Product:
    properties:
      brand:
        type: string
      buyNumber:
        type: integer
      categoryKey:
        type: string
      categoryName:
        type: string
      checkCommissionPercent:
        type: number
      commissionPercent:
        type: number
      completeSpec:
        items:
          $ref: '#/definitions/products.csType'
        type: array
      createdAt:
        type: integer
      description:
        type: string
      filterArr:
        items:
          type: string
        type: array
      imageArr:
        items:
          type: string
        type: array
      lowestPrice:
        type: integer
      mainSpecs:
        items:
          $ref: '#/definitions/products.f'
        type: array
      seenNumber:
        type: integer
      spId:
        type: string
      status:
        type: string
      tags:
        items:
          type: string
        type: array
      title:
        type: string
      variationsObj:
        $ref: '#/definitions/products.v'
    type: object
  products.advanceFilter:
    properties:
      FilterStringArr:
        items:
          $ref: '#/definitions/products.fItem'
        type: array
      brand:
        items:
          type: string
        type: array
      inStock:
        type: boolean
      priceFrom:
        type: integer
      priceTo:
        type: integer
      sort:
        type: string
    type: object
  products.colorOut:
    properties:
      main:
        $ref: '#/definitions/products.productOut'
      sub:
        items:
          $ref: '#/definitions/products.productOut'
        type: array
    type: object
  products.csType:
    properties:
      items:
        items:
          $ref: '#/definitions/products.f'
        type: array
      name:
        type: string
    type: object
  products.f:
    properties:
      name:
        type: string
      value:
        type: string
    type: object
  products.fItem:
    properties:
      name:
        type: string
      value:
        items:
          type: string
        type: array
    type: object
  products.filter:
    properties:
      brand:
        type: string
    type: object
  products.getProductByKeyResponse:
    properties:
      Product:
        $ref: '#/definitions/products.productOut'
      estelamArr:
        items:
          $ref: '#/definitions/addBuyMethod.EstelamOut'
        type: array
      priceArr:
        items:
          $ref: '#/definitions/addBuyMethod.PriceOut'
        type: array
    type: object
  products.productInfo:
    properties:
      brand:
        type: string
      categoryKey:
        type: string
      description:
        type: string
      fields:
        items:
          $ref: '#/definitions/products.f'
        type: array
      imageArr:
        items:
          type: string
        type: array
      spId:
        type: string
      tags:
        items:
          type: string
        type: array
      variationsObj:
        items:
          properties:
            description:
              type: string
            name:
              type: string
          type: object
        type: array
    type: object
  products.productOut:
    properties:
      _id:
        type: string
      _key:
        type: string
      _rev:
        type: string
      brand:
        type: string
      buyNumber:
        type: integer
      categoryKey:
        type: string
      categoryName:
        type: string
      checkCommissionPercent:
        type: number
      commissionPercent:
        type: number
      completeSpec:
        items:
          $ref: '#/definitions/products.csType'
        type: array
      createdAt:
        type: integer
      description:
        type: string
      filterArr:
        items:
          type: string
        type: array
      imageArr:
        items:
          type: string
        type: array
      lowestPrice:
        type: integer
      mainSpecs:
        items:
          $ref: '#/definitions/products.f'
        type: array
      seenNumber:
        type: integer
      spId:
        type: string
      status:
        type: string
      tags:
        items:
          type: string
        type: array
      title:
        type: string
      variationsObj:
        $ref: '#/definitions/products.v'
    type: object
  products.search:
    properties:
      searchString:
        type: string
    type: object
  products.v:
    properties:
      title:
        type: string
      variations:
        items:
          properties:
            description:
              type: string
            name:
              type: string
          type: object
        type: array
    type: object
...
info:
  contact:
    email: armansal1519@gmail.com
    name: Arman Salehi
  description: This is an API for Bamachoub Application
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: Bamachoub Application
  version: "2.0"
paths:
...
  /payment/by-check:
    post:
      consumes:
      - application/json
      description: create check by image
      parameters:
      - description: data
        in: body
        name: checkByImage
        required: true
        schema:
          $ref: '#/definitions/paymentAndWallet.checkByImage'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/paymentAndWallet.paymentOut'
        "404":
          description: Not Found
          schema:
            type: string
      summary: create check by image
      tags:
      - payment
  /payment/by-image:
    post:
      consumes:
      - application/json
      description: get payment by image
      parameters:
      - description: data
        in: body
        name: PaymentByImage
        required: true
        schema:
          $ref: '#/definitions/paymentAndWallet.PaymentByImage'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/paymentAndWallet.paymentOut'
        "404":
          description: Not Found
          schema:
            type: string
      summary: get payment by image
      tags:
      - payment
  /payment/by-url:
    post:
      consumes:
      - application/json
      description: get payment by url
      parameters:
      - description: key
        in: path
        name: key
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              type: string
            type: array
        "404":
          description: Not Found
          schema:
            type: string
      summary: get payment by url
      tags:
      - payment
  /payment/filter:
    post:
      consumes:
      - application/json
      description: fet filtered payment if filter is empty return all
      parameters:
      - description: Offset
        in: query
        name: offset
        required: true
        type: integer
      - description: limit
        in: query
        name: limit
        required: true
        type: integer
      - description: data
        in: body
        name: filter
        required: true
        schema:
          $ref: '#/definitions/paymentAndWallet.filter'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
      summary: fet filtered payment
      tags:
      - payment
  /payment/verify-check/{key}:
    post:
      consumes:
      - application/json
      description: verity check
      parameters:
      - description: key
        in: path
        name: key
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              type: string
            type: array
        "404":
          description: Not Found
          schema:
            type: string
      summary: verity check
      tags:
      - payment
  /payment/verify-image/{key}:
    post:
      consumes:
      - application/json
      description: verity payment
      parameters:
      - description: key
        in: path
        name: key
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              type: string
            type: array
        "404":
          description: Not Found
          schema:
            type: string
      summary: verity payment
      tags:
      - payment
  /payment/verify-url/{key}:
    post:
      consumes:
      - application/json
      description: verity payment by url
      parameters:
      - description: key
        in: path
        name: key
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              type: string
            type: array
        "404":
          description: Not Found
          schema:
            type: string
      summary: verity payment by url
      tags:
      - payment
...
  /products/:
    post:
      consumes:
      - application/json
      description: create products
      parameters:
      - description: data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/products.productInfo'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              type: string
            type: array
        "404":
          description: Not Found
          schema:
            type: string
      summary: create products
      tags:
      - products
  /products/{categoryurl}/{key}:
    delete:
      consumes:
      - application/json
      description: delete product by key
      parameters:
      - description: category url
        in: path
        name: categoryurl
        required: true
        type: string
      - description: Product key
        in: path
        name: key
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
      summary: deletes product
      tags:
      - products
    put:
      consumes:
      - application/json
      description: update product by key
      parameters:
      - description: category url
        in: path
        name: categoryurl
        required: true
        type: string
      - description: Product key
        in: path
        name: key
        required: true
        type: string
      - description: data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/products.Product'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
      summary: updates product
      tags:
      - products
  /products/advance-filter/{categoryurl}/{categorykey}:
    post:
      consumes:
      - application/json
      description: advance filter in products
      parameters:
      - description: category url
        in: path
        name: categoryurl
        required: true
        type: string
      - description: category key
        in: path
        name: categorykey
        required: true
        type: string
      - description: data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/products.advanceFilter'
      - description: Offset
        in: query
        name: offset
        required: true
        type: integer
      - description: limit
        in: query
        name: limit
        required: true
        type: integer
      - description: sample
        in: query
        name: sample
        required: true
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/products.productOut'
            type: array
        "404":
          description: Not Found
          schema:
            type: string
      summary: advance filter in products
      tags:
      - products
  /products/basic-filter/{categoryurl}:
    post:
      consumes:
      - application/json
      description: basic filter in products
      parameters:
      - description: category url
        in: path
        name: categoryurl
        required: true
        type: string
      - description: Offset
        in: query
        name: offset
        required: true
        type: integer
      - description: limit
        in: query
        name: limit
        required: true
        type: integer
      - description: data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/products.filter'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/products.productOut'
            type: array
        "404":
          description: Not Found
          schema:
            type: string
      summary: basic filter in products
      tags:
      - products
  /products/basic-search/{categoryurl}:
    post:
      consumes:
      - application/json
      description: basic search in products
      parameters:
      - description: category url
        in: path
        name: categoryurl
        required: true
        type: string
      - description: data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/products.search'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/products.productOut'
            type: array
        "404":
          description: Not Found
          schema:
            type: string
      summary: basic search in products
      tags:
      - products
  /products/cat/{categoryurl}/{categorykey}:
    get:
      consumes:
      - application/json
      description: return products attached to that category
      parameters:
      - description: category url
        in: path
        name: categoryurl
        required: true
        type: string
      - description: category key
        in: path
        name: categorykey
        required: true
        type: string
      - description: Offset
        in: query
        name: offset
        required: true
        type: integer
      - description: limit
        in: query
        name: limit
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/products.productOut'
            type: array
        "404":
          description: Not Found
          schema:
            type: string
      summary: return products attached to that category
      tags:
      - products
  /products/color/{spId}/{productKey}:
    post:
      consumes:
      - application/json
      description: return products with same color code
      parameters:
      - description: spId
        in: path
        name: spId
        required: true
        type: string
      - description: productKey
        in: path
        name: productKey
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/products.colorOut'
        "404":
          description: Not Found
          schema:
            type: string
      summary: return products with same color code
      tags:
      - products
  /products/one/{categoryurl}/{key}:
    get:
      consumes:
      - application/json
      description: return Product by key with estelamArr and priceArr
      parameters:
      - description: category url
        in: path
        name: categoryurl
        required: true
        type: string
      - description: Product key
        in: path
        name: key
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/products.getProductByKeyResponse'
        "404":
          description: Not Found
          schema:
            type: string
      summary: return Product by key
      tags:
      - products
  /rejection/pool:
    get:
      consumes:
      - application/json
      description: get all rejection pool items
      parameters:
      - description: Offset
        in: query
        name: offset
        required: true
        type: integer
      - description: limit
        in: query
        name: limit
        required: true
        type: integer
      - description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/paymentAndWallet.rejectionPoolItemOut'
            type: array
        "404":
          description: Not Found
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: get all rejection pool items
      tags:
      - rejection pool
...
  /suppliers-confirmation:
    get:
      consumes:
      - application/json
      description: get orders for supplier confirmation
      parameters:
      - description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/paymentAndWallet.getSupplierConfirmationResponse'
            type: array
        "404":
          description: Not Found
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: get orders for supplier confirmation
      tags:
      - supplier confirmation
  /suppliers-confirmation/approve/{infoKey}:
    post:
      consumes:
      - application/json
      description: approve order
      parameters:
      - description: infoKey
        in: path
        name: infoKey
        required: true
        type: string
      - description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: approve order
      tags:
      - supplier confirmation
  /suppliers-confirmation/reject/{infoKey}:
    post:
      consumes:
      - application/json
      description: reject order
      parameters:
      - description: infoKey
        in: path
        name: infoKey
        required: true
        type: string
      - description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: reject order
      tags:
      - supplier confirmation
...
  /wallet/data:
    get:
      consumes:
      - application/json
      description: return wallet amount , income , outcome and if withHistory query
        eq ture wallet history
      parameters:
      - description: withHistory
        in: query
        name: withHistory
        required: true
        type: boolean
      - description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/paymentAndWallet.supplierPageResponse'
        "404":
          description: Not Found
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: return wallet amount , income , outcome
      tags:
      - wallet
  /wallet/url:
    post:
      consumes:
      - application/json
      description: get payment url for wallet
      parameters:
      - description: data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/paymentAndWallet.addToWallet'
      - description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: get payment url for wallet
      tags:
      - wallet
  /wallet/verify/{key}:
    post:
      consumes:
      - application/json
      description: verify tx to wallet
      parameters:
      - description: key
        in: path
        name: key
        required: true
        type: string
      - description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: verify tx to wallet
      tags:
      - wallet
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"